Network/DevVirtioNet_1_0.cpp: Fix some Windows build warnings and check for more
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@84384 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/VirtIO/Virtio_1_0.h
@@ -100,9 +100,9 @@ typedef struct VIRTIO_DESC_CHAIN
     uint32_t            u32Magic;                                   /**< Magic value, VIRTIO_DESC_CHAIN_MAGIC.    */
     uint32_t volatile   cRefs;                                      /**< Reference counter. */
     uint32_t            uHeadIdx;                                   /**< Head idx of associated desc chain        */
-    uint32_t            cbPhysSend;                                 /**< Total size of src buffer                 */
+    size_t              cbPhysSend;                                 /**< Total size of src buffer                 */
     PVIRTIOSGBUF        pSgPhysSend;                                /**< Phys S/G/ buf for data from guest        */
-    uint32_t            cbPhysReturn;                               /**< Total size of dst buffer                 */
+    size_t              cbPhysReturn;                               /**< Total size of dst buffer                 */
     PVIRTIOSGBUF        pSgPhysReturn;                              /**< Phys S/G buf to store result for guest   */
 
     /** @name Internal (bird combined 5 allocations into a single), fingers off.
